Rename util file to firestore.ts
